Got kernel booting up to boot/head.S startup_32 which loops forever for now
Several things are going on in this commit. The objective was to get the
kernel starting up by compiling only the minimal amount of code required to
get boot/head.S startup32() executing. This involved getting the compression
code working and dealing with changes due to all binaries being ELF32 instead
of a.out. The most relevant code came from the 2.4 kernel.

Summary of the changes
- Changed tools/system target to just include head.o
- Changed tools/zSystem target to just include head.o
- Made boot/head.S loop forever
- Delete xtract piggyback, this is now done by objcopy and ld
- Added a a.out conversion to zImage target
- Updated tools/build.c to use the zSystem.out binary without expecting a a.out header.
- Added string.c and string.h from 4.5 kernel for memcmp function no longer included as a builtin.
- Added -fno-stack-protector flag so no linkage is attempted to stack protector functions
- Added temporarily hardcoded include dir -I/home/ubuntu/mac/slackware-1.0.1-include,
  so the kernel is built using the historic headers
- Updated linking & binary generation code from 2.4 kernel
- Fixed ELF style symbols in zBoot/head.S
- Fixed missing entry symbol error by adding .globl startup_32