Got init_sched() working, with quite a few compromises
What's working
- (non-fast) interrupts
- ret_from_syscall
- handle_bottom_half
- ret_from_sys_call

I stubbed a bunch of things in wip just to get everything compiling
- panic
- do_signal
- syscall_trace
- save_v86_state

Not much if this can be tested just yet. All I know is that st() works
and keyboard interrupts are still happening, so the system looks sane. More
work is needed before any of this can be tested.