Added start_kernel: fork()
Just as some work in progress & to test it, main calls test_fork() which
creates 10 processes that just print the pid and rudely call idle(). That
way it's possible to tell when the context switches happen. I also added a
little blit in the top right video memory to show when switches are happening.
Finally, error handling will put a red T or P in the corner if an unhandled
trap or page fault happens.