Fixed keyboard interrupt handler also handling mouse events
When running under qemu, this would lead to addition keyboard "7" keypresses
whenever the mouse was moved and then a key pressed. The reason for the
"7" is that the ps/2 mouse sends "08 00 00" which is interpreted as a
"7" keydown following by two keyboard errors. Mouse events are
ignored by checking bit 0x20, which indicates mouse data.

The fix isn't perfect though. It's not possible to type and move the mouse
at the same time, data is lost either in the keyboard or the mouse. But
at least now they don't interfere with each other.